+++ CHANGELOG.md
@@ -6,6 +6,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
+### Changed
+
+- Improve validation of FEP-0ea0 payment links.
+
 ## [1.23.0] - 2023-05-03
 
 ### Added